app/requirements.txt
flask==2.1.3
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
Flask-RESTful==0.3.9
Flask-Migrate==2.6.0
marshmallow==3.17.0
web3==5.30.0
PyJWT==1.7.1
apispec==5.2.2
apispec-webframeworks==0.5.2
python-dotenv==0.20.0